refactor(dashboard): extract sidebar, tabs, and workflow hooks
Phase 3 of frontend architecture refactor. Cuts Dashboard.tsx from 584
to 199 lines by extracting:

- src/lib/types.ts now holds TranscriptionInvokeResult and RecordingResult
  (previously declared inline in Dashboard).
- useRecordingWorkflow consolidates isRecording/isTranscribing state,
  the audio-captured / recording-state / recording-cancelled listeners,
  start/stop/success sound effects, transcribeAudio and
  handleToggleRecording. setSelectedTranscription is injected via an
  onTranscriptionAdded callback so the selection stays in Dashboard.
- useNotesWorkflow owns the open-tabs state for the notes modal
  (editorOpen, openNoteIds, activeNoteId) plus create/open/close/delete
  handlers.
- DashboardSidebar renders the left nav (exports DASHBOARD_NAV_ITEMS and
  the DashboardTabId union).
- HistoriqueTab is a presentational block for the recording + history
  view.
- dashboard-header.tsx moved to src/components/common/DashboardHeader.tsx.

No behavior change. Progress tracked in docs/REFACTOR_FRONTEND_PROGRESS.md.
